move RecipesBase -> weakdep
